Refactoring of load_dataset
This refactoring produces exactly the same output as the current
implementation.
The refactoring separates the code into sub functions for better
readability and easier extension.

The refactoring does not handle errors in the code. E.g. the snt_name
list is not reordered when the chunks are sorted by length. This will be
addressed in a separate commit / pull request.